+++ components/Footer/index.js
@@ -3,7 +3,6 @@
 require('./styles.css');
 
 import React from 'react';
-import {Link} from 'react-router';
 
 class Footer extends React.Component {
   render(): ?ReactElement {
